improve debuggability of could not eliminate self (#846)
This is an attempt at making it easier for people willing to change the model but not diving into thunder code to find what use of self could not be eliminated. (e.g. with #606 , you would find a chain V_0 <- 𝜙_0 <- 𝜙_4 <- V_... (self.dropout having 𝜙_4 as parent).

Of course, we might condition showing the backtrace on some env var or so.

With the new tools, I found and fixed: #606 here.